requirements.txt
aiohttp==3.7.4.post0
amqp==2.6.1
appdirs==1.4.3
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
bs4==0.0.1
CacheControl==0.12.6
celery==4.4.7
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
Django==3.2
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-redis==5.0.0
django-sslserver==0.22
django-timezone-field==4.1.2
djangorestframework==3.12.4
elasticsearch==7.13.1
flower==0.9.7
gunicorn==20.1.0
html5lib==1.0.1
humanize==3.7.0
idna==2.8
ipaddr==2.2.0
joblib==1.0.1
kombu==4.6.11
lockfile==0.12.2
msgpack==0.6.2
msoffcrypto-tool==4.11.0
multidict==5.1.0
nltk==3.6.2
oauthlib==3.1.0
olefile==0.46
packaging==20.3
pep517==0.8.2
progress==1.5
prometheus-client==0.8.0
prompt-toolkit==3.0.18
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.6
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytoml==0.1.21
pytz==2021.1
redis==3.5.3
regex==2021.4.4
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
tornado==6.1
tqdm==4.61.0
typing-extensions==3.10.0.0
urllib3==1.25.8
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
yarl==1.6.3
